EDITOR-#25. Init copy
