--- editor-server/README.md
@@ -1,9 +1,0 @@
-# Editor Server
-
-Responsible for communicating to the database.
-
-## Development
-
-```bash
-yarn start
-```
