--- editor-server/src/color-resolver.ts
@@ -1,21 +1,0 @@
-import {
-    Resolver,
-    Query,
-    FieldResolver,
-    Arg,
-    Ctx,
-    Root,
-    Mutation,
-    Int,
-    ResolverInterface,
-} from 'type-graphql';
-
-@Resolver()
-class ColorResolver {
-    @Query(returns => String)
-    async color(@Arg("color") color: string) {
-        console.log(color)
-    }
-}
-
-export default ColorResolver
