--- editor-server/src/resolvers/Mutation.ts
@@ -1,4 +1,0 @@
-const Mutation = {
-    
-}
-export { Mutation as default };
