--- editor-server/src/resolvers/Subscription.ts
@@ -1,7 +1,0 @@
-import { withFilter } from "graphql-subscriptions";
-
-const Subscription = {
-    
-};
-
-export { Subscription as default };
