--- editor-server/src/resolvers/index.ts
@@ -1,12 +1,0 @@
-import Query from "./Query"
-import Mutation from "./Mutation"
-import Subscription from "./Subscription"
-
-export default {
-    Query,
-    // Mutation,
-    Subscription
-}
-
-
-
