+++ editor-server/src/resolvers/types/color-input.ts
@@ -1,0 +1,10 @@
+import { InputType, Field } from "type-graphql";
+
+@InputType()
+export class ColorInput {
+    @Field()
+    color: string;
+
+    @Field()
+    colorCode: string
+}
